Add files via upload
The zip file contains data from analog mixed-signal circuits. there are two datasets, one is OPMAP1 and another is comparator data. both have two classes that are faulty and fault-free. there are 14 nets in the case of the OPMAP1 dataset and 20 nets in the case of the comparator dataset. the shape of OPMAP1 data is (38,14,99) and the shape of comparator data is (48,20,321). there are equal examples of faulty and fault-free in both datasets.